Implements the master-subordinate logic
Issue: #5
Implements the logic in the Python program for the communication
between the master server and the subordinates ones, via RPC.
Although the delegated tasks are basic, mimicking a time-consuming
operation, these will be extended when the extractors will be plugged
in. Besides this, the commit creates a new file structure.

Signed-off-by: iosifache <georgeiosif@live.com>