Adds the new infrastructure of the platform
Issue: #40
Updates the Docker infrastructure with a functional one, in which a
container is created for each type of server (leader, subordinate,
and predictor-collector). Besides this, it changes (again) the folder
structure to ease the sharing of the codebase and the data between
the containers and solves some small functional bugs that were
discovered when testing the functionalities on the built
infrastructure.

Signed-off-by: iosifache <georgeiosif@live.com>